+++ src/app/views/fireside/_stream-options/stream-options.ts
@@ -1,11 +1,6 @@
 import Vue from 'vue';
 import { Component, Emit, InjectReactive } from 'vue-property-decorator';
 import { stopStreaming } from '../../../../_common/fireside/rtc/producer';
-import { setRTCDesktopVolume } from '../../../../_common/fireside/rtc/rtc';
-import { setAudioPlayback } from '../../../../_common/fireside/rtc/user';
-import AppPopper from '../../../../_common/popper/popper.vue';
-import { ScrubberCallback } from '../../../../_common/slider/slider';
-import AppSlider from '../../../../_common/slider/slider.vue';
 import { AppState, AppStore } from '../../../../_common/store/app-store';
 import { AppTooltip } from '../../../../_common/tooltip/tooltip-directive';
 import {
@@ -13,11 +8,11 @@ import {
 	FiresideControllerKey,
 } from '../../../components/fireside/controller/controller';
 import { StreamSetupModal } from '../../../components/fireside/stream/setup/setup-modal.service';
+import AppFiresideSettingsPopper from '../_settings-popper/settings-popper.vue';
 
 @Component({
 	components: {
-		AppPopper,
-		AppSlider,
+		AppFiresideSettingsPopper,
 	},
 	directives: {
 		AppTooltip,
@@ -32,30 +27,6 @@ export default class AppFiresideStreamOptions extends Vue {
 	@Emit('show-popper') emitShowPopper() {}
 	@Emit('hide-popper') emitHidePopper() {}
 
-	get desktopVolume() {
-		return this.c.rtc?.desktopVolume ?? 1;
-	}
-
-	get shouldShowVolumeControls() {
-		return !!this.c.rtc?.shouldShowVolumeControls;
-	}
-
-	get shouldShowAsMuted() {
-		return this.shouldShowVolumeControls && this.desktopVolume === 0;
-	}
-
-	get shouldMute() {
-		return this.c.rtc?.users.some(i => !i.micAudioMuted) ?? false;
-	}
-
-	muteAll() {
-		return this.c.rtc?.users.forEach(i => setAudioPlayback(i, false));
-	}
-
-	unmuteAll() {
-		return this.c.rtc?.users.forEach(i => setAudioPlayback(i, true));
-	}
-
 	onClickEditStream() {
 		StreamSetupModal.show(this.c);
 	}
@@ -66,9 +37,5 @@ export default class AppFiresideStreamOptions extends Vue {
 		}
 
 		stopStreaming(this.c.rtc.producer);
-	}
-
-	onVolumeScrub({ percent }: ScrubberCallback) {
-		setRTCDesktopVolume(this.c.rtc!, percent);
 	}
 }
