+++ src/app/views/fireside/_stream/stream.vue
@@ -11,11 +11,9 @@
 		<template v-if="hasVideo">
 			<template v-if="videoPaused">
 				<transition>
-					<template v-if="!hasOverlayItems">
-						<div class="-paused-indicator -click-target anim-fade-leave-shrink">
-							<app-jolticon class="-paused-indicator-icon" icon="play" />
-						</div>
-					</template>
+					<div class="-paused-indicator -click-target anim-fade-leave-shrink">
+						<app-jolticon class="-paused-indicator-icon" icon="play" />
+					</div>
 				</transition>
 			</template>
 			<template v-else-if="isLoadingVideo">
@@ -54,36 +52,51 @@
 		>
 			<template v-if="shouldShowUI">
 				<div class="-overlay-inner">
-					<div v-if="memberCount" class="-overlay-members">
-						<translate
-							:translate-n="memberCount"
-							:translate-params="{ count: number(memberCount) }"
-							translate-plural="%{ count } members"
-						>
-							%{ count } member
-						</translate>
-					</div>
-
-					<div class="-overlay-playback">
-						<div
-							v-if="shouldShowOverlayPlayback"
-							class="-overlay-playback-inner"
-							@click="togglePlayback"
-						>
-							<app-jolticon
-								class="-paused-indicator-icon"
-								:icon="videoPaused ? 'play' : 'pause'"
-							/>
+					<div class="-overlay-top -control">
+						<div v-if="hasHeader" style="flex: auto; overflow: hidden">
+							<app-fireside-header is-overlay />
+							<div class="-overlay-members">
+								<translate
+									:translate-n="memberCount"
+									:translate-params="{ count: number(memberCount) }"
+									translate-plural="%{ count } members"
+								>
+									%{ count } member
+								</translate>
+							</div>
 						</div>
 					</div>
 
-					<div v-if="showOverlayHosts" class="-overlay-hosts -control">
-						<app-fireside-host-list
-							hide-thumb-options
-							@show-popper="onHostOptionsShow"
-							@hide-popper="onHostOptionsHide"
-						/>
+					<div class="-overlay-bottom -control">
+						<div class="-video-controls">
+							<div v-if="hasVideo">
+								<app-button
+									circle
+									trans
+									overlay
+									:icon="videoPaused ? 'play' : 'pause'"
+									@click.capture.stop="togglePlayback"
+								/>
+							</div>
+
+							<div v-if="hasVolumeControls" class="-volume">
+								<app-jolticon icon="audio" />
+								<app-slider
+									class="-volume-slider"
+									:percent="desktopVolume"
+									@scrub="onVolumeScrub"
+								/>
+							</div>
+						</div>
 					</div>
+
+					<app-fireside-host-list
+						v-if="hasHosts"
+						class="-hosts"
+						hide-thumb-options
+						@show-popper="onHostOptionsShow"
+						@hide-popper="onHostOptionsHide"
+					/>
 				</div>
 			</template>
 		</div>
@@ -120,6 +133,14 @@
 @import '~styles/variables'
 @import '~styles-lib/mixins'
 
+$-text-shadow = 1px 1px 3px rgba($black, 0.5)
+$-z-overlay = 1
+$-z-combo = 2
+$-z-paused = 2
+
+.jolticon
+	text-shadow: $-text-shadow
+
 .-stream
 .-video-player
 	&
@@ -130,10 +151,10 @@
 		bottom: 0
 		left: 0
 		color: var(--theme-fg)
-		text-shadow: 1px 1px 3px rgba($black, 0.5)
+		text-shadow: $-text-shadow
 
 	> .-overlay
-		z-index: 1
+		z-index: $-z-overlay
 		opacity: 0
 		transition: all 200ms $strong-ease-out
 
@@ -149,10 +170,7 @@
 	width: 100%
 	display: flex
 	flex-direction: column
-	padding: 8px 0
-
-	> *
-		flex: 1
+	padding: 8px
 
 .-visible-center
 	opacity: 1 !important
@@ -173,21 +191,15 @@
 	left: 0
 
 .-overlay-members
-	padding: 0 8px
+	opacity: 0.75
 	font-weight: bold
 
-.-overlay-playback
-	flex: auto
+.-overlay-top
 	display: flex
-	align-items: center
-	justify-content: center
-	min-height: 0
+	align-items: flex-start
+	margin-bottom: auto
 
-	&-inner
-		padding: 16px
-		margin: -(@padding)
-
-.-overlay-hosts
+.-overlay-bottom
 	display: flex
 	align-items: flex-end
 
@@ -205,9 +217,29 @@
 	display: flex
 	align-items: center
 	justify-content: center
+	z-index: $-z-paused
+	pointer-events: none
 
 	&-icon
 		font-size: 60px
+
+.-video-controls
+	display: flex
+	align-items: center
+	flex: 1
+	grid-gap: 12px
+
+	.-volume
+		display: inline-flex
+		align-items: center
+		flex: auto
+		grid-gap: 4px
+
+		&-slider
+			max-width: 200px
+
+.-hosts
+	margin-top: 8px
 
 .-combo
 	position: absolute
@@ -218,7 +250,7 @@
 	font-weight: bold
 	color: white
 	align-items: center
-	z-index: 2
+	z-index: $-z-combo
 	transition: opacity 200ms $strong-ease-out
 
 	&.-fade
